notification-backend/requirements.txt
alembic==1.4.2
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
dataclasses==0.6
fastapi==0.61.1
fastapi-jwt-auth==0.1.0
FastAPI-SQLAlchemy==0.1.0
flake8==3.8.3
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
sqlalchemy-seed==0.3.0
starlette==0.13.6
toml==0.10.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
zipp==3.1.0
